added mappings to the SemanticDimensionsMapper, added tests

+++ languages/java-cpg/src/main/java/de/jplag/java_cpg/token/semantic/SemanticCpgTokenEquivalenceModel.java
@@ -8,7 +8,7 @@ public class SemanticCpgTokenEquivalenceModel extends CpgTokenEquivalenceModel {
     private final double semanticThreshold;
 
     public SemanticCpgTokenEquivalenceModel(int semanticThreshold) {
-        this.semanticThreshold = (double) semanticThreshold;
+        this.semanticThreshold = semanticThreshold;
     }
 
     @Override
